requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cramjam==2.7.0
cryptography==40.0.2
cycler==0.11.0
et-xmlfile==1.1.0
fastparquet==2023.8.0
fonttools==4.39.3
fpdf==1.7.2
frozendict==2.3.7
fsspec==2023.10.0
html5lib==1.1
idna==3.4
install==1.3.5
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
monthly-returns-heatmap==0.0.11
mplcyberpunk==0.7.0
multitasking==0.0.11
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.0
patsy==0.5.3
Pillow==9.5.0
pyarrow==13.0.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
QuantStats==0.0.59
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
statsmodels==0.14.0
tabulate==0.9.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
yfinance==0.2.18
